+++ lib/membrane/core/element/demand_controller.ex
@@ -101,7 +101,12 @@ defmodule Membrane.Core.Element.DemandController do
   @spec decrease_demand_by_outgoing_buffers(Pad.ref(), [Buffer.t()], State.t()) :: State.t()
   def decrease_demand_by_outgoing_buffers(pad_ref, buffers, state) do
     pad_data = PadModel.get_data!(state, pad_ref)
-    buffers_size = Buffer.Metric.from_unit(pad_data.demand_unit).buffers_size(buffers)
+
+    {:ok, buffers_size} =
+      Membrane.Core.Element.ManualFlowController.BufferMetric.buffers_size(
+        pad_data.demand_unit,
+        buffers
+      )
 
     demand = pad_data.demand - buffers_size
     {decrease_result, atomic_demand} = AtomicDemand.decrease(pad_data.atomic_demand, buffers_size)
@@ -115,7 +120,8 @@ defmodule Membrane.Core.Element.DemandController do
     |> PadModel.set_data!(pad_ref, %{
       pad_data
       | demand: demand,
-        atomic_demand: atomic_demand
+        atomic_demand: atomic_demand,
+        uninterrupted_redemands: 0
     })
   end